Fix e2e
