+++ api/src/data/connection.ts
@@ -1,0 +1,19 @@
+function requireEnv(name: string): string {
+	const value = process.env[name];
+	if (!value) {
+		throw new Error(`Missing required database env: ${name}`);
+	}
+	return value;
+}
+
+export function getConnectionUrl(): string {
+	const user = requireEnv("DB_USER");
+	const pass = requireEnv("DB_PASS");
+	const port = requireEnv("DB_PORT");
+	const domain = requireEnv("DB_DOMAIN");
+	const database = requireEnv("DB_EVY_DATABASE");
+
+	const encodedUser = encodeURIComponent(user);
+	const encodedPass = encodeURIComponent(pass);
+	return `postgresql://${encodedUser}:${encodedPass}@${domain}:${port}/${database}`;
+}
