+++ api/src/database/db.ts
@@ -2,26 +2,7 @@ import { SQL } from "bun";
 import { drizzle } from "drizzle-orm/bun-sql";
 
 import * as schema from "../../../types/generated/ts/db/schema.generated";
-
-function requireEnv(name: string): string {
-	const value = process.env[name];
-	if (!value) {
-		throw new Error(`Missing required database env: ${name}`);
-	}
-	return value;
-}
-
-function getConnectionUrl(): string {
-	const user = requireEnv("DB_USER");
-	const pass = requireEnv("DB_PASS");
-	const port = requireEnv("DB_PORT");
-	const domain = requireEnv("DB_DOMAIN");
-	const database = requireEnv("DB_EVY_DATABASE");
-
-	const encodedUser = encodeURIComponent(user);
-	const encodedPass = encodeURIComponent(pass);
-	return `postgresql://${encodedUser}:${encodedPass}@${domain}:${port}/${database}`;
-}
+import { getConnectionUrl } from "../data/connection";
 
 export type EvyDb = ReturnType<typeof createDb>;
